Shuffling around the steps
We now have the new `Opening` namespace or group, and it will contain apart from the postcode question, a couple more questions that previously were in the `MIAM` namespace/group.

Updated the `OpeningDecisionTree` and the `MiamDecisionTree` to take this changes into consideration.
