--- app/controllers/steps/miam/child_protection_info_controller.rb
@@ -1,7 +1,0 @@
-module Steps
-  module Miam
-    class ChildProtectionInfoController < Steps::MiamStepController
-      def show; end
-    end
-  end
-end
