--- app/controllers/steps/miam/consent_order_sought_controller.rb
@@ -1,7 +1,0 @@
-module Steps
-  module Miam
-    class ConsentOrderSoughtController < Steps::MiamStepController
-      def show; end
-    end
-  end
-end
