+++ app/controllers/steps/opening/child_protection_cases_controller.rb
@@ -1,6 +1,6 @@
 module Steps
-  module Miam
-    class ChildProtectionCasesController < Steps::MiamStepController
+  module Opening
+    class ChildProtectionCasesController < Steps::OpeningStepController
       def edit
         @form_object = ChildProtectionCasesForm.new(
           c100_application: current_c100_application,
