+++ app/controllers/steps/opening/child_protection_info_controller.rb
@@ -1,0 +1,7 @@
+module Steps
+  module Opening
+    class ChildProtectionInfoController < Steps::OpeningStepController
+      def show; end
+    end
+  end
+end
