+++ app/controllers/steps/opening/consent_order_sought_controller.rb
@@ -1,0 +1,7 @@
+module Steps
+  module Opening
+    class ConsentOrderSoughtController < Steps::OpeningStepController
+      def show; end
+    end
+  end
+end
