+++ app/forms/steps/opening/child_protection_cases_form.rb
@@ -1,5 +1,5 @@
 module Steps
-  module Miam
+  module Opening
     class ChildProtectionCasesForm < BaseForm
       include SingleQuestionForm
