+++ app/views/entrypoint/how_long.en.html.erb
@@ -23,7 +23,7 @@
       delete any information after this time.
     </p>
 
-    <%= link_button 'Continue', edit_steps_miam_consent_order_path,
+    <%= link_button 'Continue', edit_steps_opening_postcode_path,
                     class: 'app-button--l app-button--inverted govuk-!-margin-top-3 govuk-!-margin-bottom-1' %>
   </div>
 </div>
