+++ app/views/steps/opening/consent_order_sought/show.en.html.erb
@@ -13,6 +13,6 @@
       You and the other person must sign this agreement.
     </p>
 
-    <%= link_button 'Continue', edit_steps_miam_child_protection_cases_path %>
+    <%= link_button 'Continue', edit_steps_opening_child_protection_cases_path %>
   </div>
 </div>
