+++ app/views/steps/opening/postcode/edit.html.erb
@@ -1,5 +1,5 @@
 <% title t('.page_title') %>
-<% step_header %>
+<% step_header(path: entrypoint_how_long_path) %>
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
