+++ config/routes.rb
@@ -113,6 +113,10 @@ Rails.application.routes.draw do
       edit_step :postcode
       show_step :error_but_continue
       show_step :no_court_found
+      edit_step :consent_order
+      show_step :consent_order_sought
+      edit_step :child_protection_cases
+      show_step :child_protection_info
     end
     namespace :application do
       edit_step :previous_proceedings
@@ -139,10 +143,6 @@ Rails.application.routes.draw do
       edit_step :protection
     end
     namespace :miam do
-      edit_step :consent_order
-      show_step :consent_order_sought
-      edit_step :child_protection_cases
-      show_step :child_protection_info
       edit_step :acknowledgement
       edit_step :attended
       edit_step :exemption_claim
