--- spec/controllers/steps/miam/child_protection_cases_controller_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe Steps::Miam::ChildProtectionCasesController, type: :controller do
-  it_behaves_like 'an intermediate step controller', Steps::Miam::ChildProtectionCasesForm, C100App::MiamDecisionTree
-end
