--- spec/controllers/steps/miam/child_protection_info_controller_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe Steps::Miam::ChildProtectionInfoController, type: :controller do
-  it_behaves_like 'a show step controller'
-end
