--- spec/controllers/steps/miam/consent_order_sought_controller_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe Steps::Miam::ConsentOrderSoughtController, type: :controller do
-  it_behaves_like 'a show step controller'
-end
