+++ spec/controllers/steps/opening/child_protection_cases_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Steps::Opening::ChildProtectionCasesController, type: :controller do
+  it_behaves_like 'an intermediate step controller', Steps::Opening::ChildProtectionCasesForm, C100App::OpeningDecisionTree
+end
