+++ spec/controllers/steps/opening/consent_order_sought_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Steps::Opening::ConsentOrderSoughtController, type: :controller do
+  it_behaves_like 'a show step controller'
+end
