+++ spec/forms/steps/opening/child_protection_cases_form_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-RSpec.describe Steps::Miam::ChildProtectionCasesForm do
+RSpec.describe Steps::Opening::ChildProtectionCasesForm do
   it_behaves_like 'a yes-no question form',
                   attribute_name: :child_protection_cases,
                   reset_when_yes: [
